pyproject.toml
[project]
name = "dazbo-portfolio"
version = "0.1.0"
description = "Portfolio site for Dazbo, with agentic features"
authors = [
    {name = "Darren (Dazbo) Lester", email = "derailed.dash@gmail.com"},
]
dependencies = [
    "google-adk>=1.15.0,<2.0.0",
    "opentelemetry-instrumentation-google-genai>=0.1.0,<1.0.0",
    "gcsfs>=2024.11.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "google-cloud-aiplatform[evaluation]>=1.118.0,<2.0.0",
    "fastapi~=0.115.8",
    "uvicorn~=0.34.0",
    "asyncpg>=0.30.0,<1.0.0",
    "google-cloud-firestore>=2.23.0",
]
requires-python = ">=3.12,<3.14"


[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
    "nest-asyncio>=1.6.0,<2.0.0",
]

[project.optional-dependencies]
jupyter = [
    "jupyter>=1.0.0,<2.0.0",
]
lint = [
    "ruff>=0.4.6,<1.0.0",
    "ty>=0.0.1a0",
    "codespell>=2.2.0,<3.0.0",
]

[tool.ruff]
line-length = 125
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901", "B006", "B008"] # ignore line too long, too complex, FastAPI Depends

[tool.ruff.lint.isort]
known-first-party = ["app", "frontend"]

[tool.ty]
# ty is Astral's Rust-based type checker (same team as ruff/uv)
# See: https://docs.astral.sh/ty/

[tool.ty.environment]
python-version = "3.12"

[tool.ty.src]
exclude = [".venv/**", "**/*.ipynb"]

[tool.ty.rules]
# Ignore common issues with third-party libraries and dynamic code patterns
unresolved-import = "ignore"
unresolved-attribute = "ignore"
invalid-argument-type = "ignore"
invalid-assignment = "ignore"
invalid-return-type = "ignore"
possibly-missing-attribute = "ignore"

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    # Ignore Pydantic V2 deprecation warnings from third-party libs like litellm
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
    # Ignore ADK experimental warnings
    "ignore:\\[EXPERIMENTAL\\]:UserWarning:google.adk.*",
    # Ignore internal DeprecationWarnings from ADK and GenAI
    "ignore:deprecated:DeprecationWarning:google.adk.*",
    "ignore:Inheritance class AiohttpClientSession.*:DeprecationWarning:google.genai.*",
]

[tool.hatch.build.targets.wheel]
packages = ["app","frontend"]

[tool.agent-starter-pack]
# Generation metadata - enables CLI commands to understand project context
# These fields allow any generated project to work as a remote template
name = "dazbo-portfolio"
description = "Simple ReAct agent"
base_template = "adk"
agent_directory = "app"

# Generation context
generated_at = "2026-01-16T23:32:37.537374+00:00"
asp_version = "0.31.2"
